Funcionalidades: Implementação da funcionalidade de Check-in e serviços relacionados
- Adicionado o CheckInRepository para gerenciar registros de check-in.

- Criado o CheckInServic para lidar com a lógica de negócios dos check-ins.

- Introduzido o CheckInController para endpoints da API relacionados a check-ins.

- Atualizado o ColaboradorRepository para incluir métodos de busca por nome e filtragem por empresa.

- Adicionado o EmpresaRepository para gerenciar dados de empresas.

- Criado o EmpresaService para lidar com operações relacionadas a empresas.

- Implementado o AuthService  para autenticação e registro de usuários.

- Introduzido o TokenServices para geração e validação de tokens JWT.

- Adicionado o ColaboradorService para gerenciar dados e operações de colaboradores.

- Criado o UsuarioService para gerenciamento de usuários, incluindo atualizações de credenciais.

- Atualizadas as propriedades do aplicativo para incluir o JWT scret.

- Adicionados novos DTOs para respostas de CheckIn, Colaborador e Token.

- Criadas classes de entidade para CheckIn, Empresa e outras entidades necessárias.

- Atualizada a configuração de segurança para suportar autenticação JWT.

+++ empresa/src/main/java/com/empresa/empresa/Repositories/DepartamentoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.empresa.empresa.Entities.Departamento;
 
-public interface DepartamentoRepository extends JpaRepository<Departamento, Integer>{
+public interface DepartamentoRepository extends JpaRepository<Departamento, Long>{
 
 }
